Simplify existing jelly state machine
Also fixed a bug with drag and drop
