--- Assets/Scripts/Jellies/Behaviors.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 670a01c3ef674af2b6845e31174ef311
-timeCreated: 1682521322
